Split nav to own file
